requirements.txt
torch
torch-geometric
sentence-transformers
chromadb
streamlit
pypdf
python-docx
pandas
scikit-learn
reportlab
tqdm
requests
google.generativeai